requirements.txt
asgiref==3.4.1
attrs==21.2.0
awscli==1.20.46
bcrypt==3.2.0
black==22.1.0
boto3==1.18.46
botocore==1.21.46
certifi==2021.5.30
cffi==1.14.6
click==8.0.4
colorama==0.4.3
cryptography==3.4.8
Django==3.2.6
django-cors-headers==3.10.0
django-debug-toolbar==3.2.4
djangorestframework==3.12.4
docutils==0.15.2
flake8==4.0.1
gunicorn==20.1.0
iniconfig==1.1.1
jmespath==0.10.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
packaging==21.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.5.1
pluggy==1.0.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.7
